STYLE: naming convention

+++ kcu-hack/util.h
@@ -8,6 +8,6 @@
 
 DWORD getProcessID(const std::wstring& processName);
 DWORD check_pid(const std::wstring& process_name);
-uint32_t get_base_address(DWORD pid, const std::wstring& moduleName);
+uintptr_t get_base_address(DWORD pid, const std::wstring& moduleName);
 
 #endif // UTIL_H
